+++ app/controllers/user_controller.rb
@@ -1,6 +1,7 @@
 class UserController < ApplicationController
   include ApplicationHelper
   include HistoryHelper
+  include Bookmarks
 
   before_action :authenticate_user!
   before_action :set_breadcrumb
@@ -51,25 +52,9 @@ class UserController < ApplicationController
     end
 
     @bookmarks = bookmarks.map { |bookmark| BookmarkPresenter.new(bookmark) }
+    @bookmark_lists = current_user.bookmark_lists
 
-    @categories = all_bookmarks.map(&:product)
-                               .flat_map(&:sub_categories)
-                               .sort_by(&:name)
-                               .uniq
-                               .group_by(&:category)
-                               .sort_by { |category| category[0].order }
-                               .map do |c|
-                                 [
-                                   c[0],
-                                   c[1].map do |sub_category|
-                                     {
-                                       name: sub_category.name,
-                                       friendly_id: sub_category.friendly_id,
-                                       path: dashboard_bookmarks_path(category: sub_category.friendly_id)
-                                     }
-                                   end
-                                 ]
-                               end
+    @categories = get_grouped_sub_categories(bookmarks: all_bookmarks)
   end
 
   def contributions