+++ app/models/bookmark.rb
@@ -2,12 +2,14 @@ class Bookmark < ApplicationRecord
   belongs_to :user
   belongs_to :product
   belongs_to :product_variant, optional: true
+  belongs_to :bookmark_list, optional: true
 
   validates :product_id, uniqueness: { scope: [:user_id, :product_variant_id] }
 
   # :nocov:
   def self.ransackable_attributes(_auth_object = nil)
     %w[
+      bookmark_list_id_eq
       created_at
       id
       id_value