+++ db/migrate/20241004192630_add_bookmark_list_to_bookmarks.rb
@@ -1,0 +1,5 @@
+class AddBookmarkListToBookmarks < ActiveRecord::Migration[7.2]
+  def change
+    add_column :bookmarks, :bookmark_list_id, :bigint
+  end
+end