+++ db/migrate/20241005080522_add_index_to_bookmark_lists.rb
@@ -1,0 +1,5 @@
+class AddIndexToBookmarkLists < ActiveRecord::Migration[7.2]
+  def change
+    add_index :bookmark_lists, [:name, :user_id], unique: true
+  end
+end