+++ db/migrate/20241005081316_add_foreign_keys_to_bookmark_lists.rb
@@ -1,0 +1,5 @@
+class AddForeignKeysToBookmarkLists < ActiveRecord::Migration[7.2]
+  def change
+    add_foreign_key :bookmark_lists, :users
+  end
+end