+++ db/migrate/20241005082118_add_index_to_user_bookmark_lists.rb
@@ -1,0 +1,5 @@
+class AddIndexToUserBookmarkLists < ActiveRecord::Migration[7.2]
+  def change
+    add_index :bookmark_lists, :user_id
+  end
+end