+++ db/migrate/20241005082341_add_index_to_bookmarks_on_bookmark_lists.rb
@@ -1,0 +1,5 @@
+class AddIndexToBookmarksOnBookmarkLists < ActiveRecord::Migration[7.2]
+  def change
+    add_index :bookmarks, :bookmark_list_id
+  end
+end